+++ projects/np-ui-lib/CHANGELOG.md
@@ -5,6 +5,7 @@
 
 # 11.1.5
 - Add sidepanel service
+- Add new dateClass property to date picker
 
 # 11.1.4
 - Resolved css issue in datagrid